Bugfix: remove invisible carriage return in sketch.bash shebang
While trying to build on a Linux system, you would get a weird error
when trying to execute "make sketch" saying "#!/bin/bash^M bad
interpreter: file not found". This error was caused by a carriage return
character after the shebang of the script which caused the desired bash
executable to be impossible to find. This commit simply fixes the bug by
removing the unwanted CR.